Refactor: 경로 조회 문서 생성을 위한 경로 조회 요청 Steps 추가 - RequestSpecification 을 TC 에서 설정하고, Steps 의 메서드 인자로 전달
