+++ src/test/java/nextstep/subway/documentation/PathDocumentation.java
@@ -1,5 +1,7 @@
 package nextstep.subway.documentation;
 
+import static nextstep.subway.documentation.steps.PathDocumentationSteps.경로_조회_요청;
+import static nextstep.subway.documentation.steps.PathDocumentationSteps.경로_조회_요청_문서_데이터_생성;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
@@ -9,41 +11,18 @@ import static org.springframework.restdocs.request.RequestDocumentation.paramete
 import static org.springframework.restdocs.request.RequestDocumentation.requestParameters;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
-import io.restassured.RestAssured;
-import java.util.List;
-import nextstep.subway.controller.PathController;
-import nextstep.subway.dto.PathRequest;
-import nextstep.subway.dto.PathResponse;
-import nextstep.subway.dto.StationResponse;
+import nextstep.subway.documentation.steps.PathDocumentationSteps.PathInformation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 
 @DisplayName("경로 조회 API 문서")
 public class PathDocumentation extends Documentation {
-
-    @MockBean
-    private PathController pathController;
 
     @Test
     @DisplayName("[성공] 경로 조회 요청 문서")
     void 경로_조회_요청_문서() {
         // Given
-        Long 교대역 = 1L;
-        Long 양재역 = 3L;
-        PathResponse 경로_조회_응답 = PathResponse.builder()
-            .stations(List.of(
-                new StationResponse(교대역, "교대역"),
-                new StationResponse(2L, "남부터미널역"),
-                new StationResponse(양재역, "양재역")
-            ))
-            .distance(5L)
-            .build();
-        Mockito.when(pathController.findPath(Mockito.any(PathRequest.class)))
-            .thenReturn(ResponseEntity.ok(경로_조회_응답));
+        PathInformation 경로 = 경로_조회_요청_문서_데이터_생성();
 
         // When
         spec.filter(document("path",
@@ -60,13 +39,7 @@ public class PathDocumentation extends Documentation {
                 fieldWithPath("distance").description("경로의 총 거리")
             )));
 
-        RestAssured
-            .given(spec)
-            .accept(MediaType.APPLICATION_JSON_VALUE)
-            .queryParam("source", 교대역)
-            .queryParam("target", 양재역)
-            .when().get("/paths")
-            .then().log().all().extract();
+        경로_조회_요청(spec, 경로.출발역, 경로.도착역);
     }
 
 }
